chore: @node/types -> 16.x

+++ packages/e2e-test/src/commands/run.ts
@@ -372,8 +372,9 @@ async function testAppServe(pluginName: string, appDir: string) {
     }
   } finally {
     // Kill entire process group, otherwise we'll end up with hanging serve processes
-    await new Promise<void>((res, rej) =>
-      killTree(startApp.pid, err => (err ? rej(err) : res())),
+    await new Promise<void>(
+      (res, rej) =>
+        startApp.pid && killTree(startApp.pid, err => (err ? rej(err) : res())),
     );
   }
 
@@ -477,8 +478,9 @@ async function testBackendStart(appDir: string, isPostgres: boolean) {
   } finally {
     print('Stopping the child process');
     // Kill entire process group, otherwise we'll end up with hanging serve processes
-    await new Promise<void>((res, rej) =>
-      killTree(child.pid, err => (err ? rej(err) : res())),
+    await new Promise<void>(
+      (res, rej) =>
+        child.pid && killTree(child.pid, err => (err ? rej(err) : res())),
     );
   }
